Feature/playwright routing (#3115)
* Create routing.spec.ts

* Delete routing.cy.js
